requirements.txt
Pillow
docopt==0.6.2